Improved dashboard controller and router. removed console.log()s.
