+++ server.js
@@ -4,12 +4,12 @@ import memorystore from 'memorystore';
 import ejs from 'ejs';
 import Docker from 'dockerode';
 import { router } from './router/index.js';
-import { sequelize, Container } from './database/models.js';
+import { sequelize } from './database/models.js';
 import { currentLoad, mem, networkStats, fsSize } from 'systeminformation';
-import { containerCard } from './components/containerCard.js';
-export var docker = new Docker();
+import { sendCheck } from './controllers/dashboard.js';
 
-export { setEvent, sse, cpu, ram, tx, rx, disk }
+export var docker = new Docker();
+export { setEvent, cpu, ram, tx, rx, disk }
 
 const app = express();
 const MemoryStore = memorystore(session);
@@ -54,14 +54,11 @@ app.listen(port, async () => {
 });
 
 let [ cpu, ram, tx, rx, disk ] = [0, 0, 0, 0, 0];
-let [ hidden, cardList, sentList ] = ['', '', ''];
-let event = false;
-let sse = false;
-let eventInfo = '';
+let [ event, sse, eventType ] = [false, false, ''];
 
 function setEvent(value, type) {
     event = value;
-    eventInfo = type;
+    eventType = type;
 }
 
 // Server metrics
@@ -82,86 +79,23 @@ let serverMetrics = async () => {
 }
 setInterval(serverMetrics, 1000);
 
-// Get hidden containers
-async function getHidden() {
-    hidden = await Container.findAll({ where: {visibility:false}});
-    hidden = hidden.map((container) => container.name);
-}
-
-// Create list of docker containers cards
-let containerCards = async () => {
-    let list = '';
-    const allContainers = await docker.listContainers({ all: true });
-    for (const container of allContainers) {
-        if (!hidden.includes(container.Names[0].slice(1))) {
-
-            let imageVersion = container.Image.split('/');
-            let service = imageVersion[imageVersion.length - 1].split(':')[0];
-            let containerId = docker.getContainer(container.Id);
-            let containerInfo = await containerId.inspect();
-            let ports_list = [];
-            try {
-            for (const [key, value] of Object.entries(containerInfo.HostConfig.PortBindings)) {
-                let ports = {
-                    check: 'checked',
-                    external: value[0].HostPort,
-                    internal: key.split('/')[0],
-                    protocol: key.split('/')[1]
-                }
-                ports_list.push(ports);
-            }
-            } catch {}
-
-            let external_port = ports_list[0]?.external || 0;
-            let internal_port = ports_list[0]?.internal || 0;
-
-            let container_info = {
-                name: container.Names[0].slice(1),
-                service: service,
-                id: container.Id,
-                state: container.State,
-                image: container.Image,
-                external_port: external_port,
-                internal_port: internal_port,
-                ports: ports_list,
-                link: 'localhost',
-            }
-            let card = containerCard(container_info);
-            list += card;
-        }
-    }
-    cardList = list;
-}
-
 
 // Docker events
 docker.getEvents((err, stream) => {
     if (err) throw err;
     stream.on('data', (chunk) => {
         event = true;
-        eventInfo = 'docker';
+        eventType = 'docker';
     });
 });
 
 // Check if the container cards need to be updated
 setInterval(async () => {
     if (event == false) { return; }
-    await getHidden();
-    await containerCards();
-    if (cardList != sentList) {
-        cardList = sentList;
-        sse = true;
-    }
+    sse = await sendCheck();
     event = false;
-}, 1000);
+}, 500);
 
-// Gets called at load and after server-side events
-router.get('/containers', async (req, res) => {
-    await getHidden();
-    await containerCards();
-    sentList = cardList;
-    res.send(cardList);
-});
 
 
 router.get('/sse_event', (req, res) => {
@@ -169,14 +103,11 @@ router.get('/sse_event', (req, res) => {
     let eventCheck = setInterval(async () => {
         if (sse == true) {
             sse = false;
-            console.log(`event: ${eventInfo}`);
-            res.write(`event: ${eventInfo}\n`);
+            res.write(`event: ${eventType}\n`);
             res.write(`data: there was an event!\n\n`);
         }
-    }, 1000);
+    }, 500);
     req.on('close', () => {
         clearInterval(eventCheck);
     });
 });
-
-
